ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
pydantic==2.5.3
python-dotenv==1.0.0
Pillow>=10.0.0
